requirements.txt
langchain
langchain-community
langchain-openai
langgraph
python-dotenv
langchain-google-genai